python/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Whisper for transcription
faster-whisper>=1.0.0

# Database
psycopg2-binary>=2.9.9

# Optional: For CUDA GPU acceleration (uncomment if you have NVIDIA GPU)
# ctranslate2>=4.0.0
